move everything under the package mod
