+++ hermes/bin/src/lib.rs
@@ -11,8 +11,6 @@ pub mod packaging;
 pub mod reactor;
 pub mod runtime_context;
 pub mod runtime_extensions;
-pub mod schema_validation;
-pub mod sign;
 pub mod wasm;
 
 #[cfg(feature = "bench")]
