+++ hermes/bin/src/main.rs
@@ -9,8 +9,6 @@ mod packaging;
 mod reactor;
 mod runtime_context;
 mod runtime_extensions;
-mod schema_validation;
-mod sign;
 mod wasm;
 
 #[cfg(feature = "bench")]
