+++ hermes/bin/src/packaging/sign/hash.rs
